ConnectToMySQL-ConnectToArduino
Sign in after filling in the credentials in the AAU MySQL Authentification.
Send all the data  through the dictionary named "Logs" created in the script "LogginManager" to the Database.
Connect to Arduino by entring the address mail, indicating the port COM we're connected to.
Some errors fixed about the positions (X and Y).