Dictionary changes
Previously, the dictionary was created each time the test is played, fixed that by moving the  dictionary  into the Awake() function to avoid this issue and after clear the list of it.

Creating a new function named sendLogs() which enables to send data after the test is done and avoid several upload calls, which allows us to minimize traffic.

Fixed one issue according the Arduino in the main scene, I had to delete the GameObject named ConnectToArduino within this scene.