Add TargetsDistance Field
Logs : Addition of the TargetsDistance field on the DB,  which contains the  values of the target size of the three game mode (Tunnel-Fitts and Goal) selected and use it to visualize it in terms of human performance on R Shiny app.
Some bugs are fixed, regarding the selection of game type before the game is launched by sending empty log and returned an error message on Unity.